Environment converted to a struct
Environment has been converted to a struct with the relevant environment properties.

As we were only using the enum to get access to relevant environment properties a struct fits better for our current use case.

The production & sandbox environments has been defined as static properties on the environment itself, giving us the same nice syntax as the caller when defining what environment to use.