Adjust how device info is handled
This change moves the device out of the view model and handles it in the networking layer instead.

This allows us to remove the global device function and simplify networking in the view model.
Furthermore this allows the device detection to be done once and the result stored instead of previously where the device would be recomputed every time a network call was made.